requirements.txt
# boto3
# fastapi
pydantic==1.10.9
# gunicorn
# uvicorn
loguru
httpx
asyncpg
sqlalchemy[asyncio]
# sqlalchemy
pgvector
asyncio
psycopg2-binary
jupyter
ipykernel
pandas
python-dotenv
openai==0.28.1 # 1.1.1 not compatible with guardrails-ai
num2words
matplotlib
plotly
scipy
scikit-learn
tiktoken
guardrails-ai
